Correct server number error

+++ mmn.h
@@ -49,7 +49,7 @@ void run_service(_system s);
 
 /* Global variables */
 double lambda_1, lamda_2, mu;			// Parameters for generate exp rv
-int samples;							// Number of events
+double samples;							// Number of events
 int count;
 queue *q_system;						// Queue which stores all events
